Fix auth and name error
