+++ app/controllers/api/v1/api_controller.rb
@@ -5,7 +5,7 @@ class Api::V1::ApiController < ActionController::API
 
   def authorize_request
     header = request.headers['Authorization']
-    header = header.split(' ').last if header
+    header = header.split.last if header
     begin
       @decoded = JsonWebToken.decode(header)
       @current_user = User.find(@decoded[:user_id])
